moved config func
